Disable automatic loading of executable and keep the runtime env active.
From the documentation [1]:

noInitialRun: If set to true, main() will not be called. The program will
still call global initializers, set up memory initialization, and so forth.
You can then call main() yourself later.

noExitRuntime: If set to true, the runtime is not shut down after run is
called. Shutting down the runtime calls shutdown callbacks, for example
atexit calls. If you want to be able to continue to use the code after
run finishes, it is safer to set this.

$ git shortlog 629744894..7ef3b940
Jean-Christophe Fillion-Robin (2):
      Force gzip'ing of js file even if it already exists.
      Associate a "pre" javascript file to customize generated js files.

[1] https://github.com/kripken/emscripten/wiki/Interacting-with-code