cmake: CMake'ified LibXml2 to support building shared lib
This is needed to avoid duplicated symbol issue when enabling xml2dsr
application. See #23

Indeed, trying to configure with "--disable-static --enable-shared" was
failing and attempting to patch the build system was not straightforward.

This commit adds a simply CMakeLists.txt along with pre-configured headers
config.h and xmlversion.h suitable for emscripten. These headers have first
been obtained by configuring the project with:

 emconfigure ./configure  \
  --without-python \
  --without-iconv \
  --without-http \
  --without-ftp \
  --disable-shared --enable-static